+++ data/examples/declaration/value/function/if-with-comment-out.hs
@@ -3,6 +3,12 @@ foo =
     then -- then comment
       undefined
     else -- else comment
-
     do
       undefined
+
+x =
+  if undefined
+    then -- then
+      ()
+    else -- else
+      ()