+++ data/examples/declaration/value/function/if-with-comment.hs
@@ -6,3 +6,12 @@ foo =
     else
       do
         undefined
+
+x =
+  if undefined
+    -- then
+    then
+      ()
+    -- else
+    else
+      ()