+++ default.nix
@@ -38,6 +38,7 @@ let
     "pipes"
     "postgrest"
     "purescript"
+    "stack"
   ];
   ormolizedPackages = doCheck:
     pkgs.lib.mapAttrs (name: p: ormolize {